PLG-313: move category form normalization in a dedicated service
