+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/forms.yml
@@ -1,0 +1,2 @@
+services:
+  Akeneo\Pim\Enrichment\Bundle\Form\CategoryFormViewNormalizer: ~
